+++ examples/realworld.svelte.dev/src/routes/auth/login.js
@@ -1,12 +1,11 @@
 import * as api from '$common/api.js';
 
-export function post(req, res) {
-	const user = req.body;
-
-	api.post('users/login', { user }).then(response => {
-		if (response.user) req.session.user = response.user;
-		res.setHeader('Content-Type', 'application/json');
-
-		res.end(JSON.stringify(response));
+export async function post({ body }) {
+	const response = await api.post('users/login', {
+		user: body
 	});
+
+	return {
+		body: response
+	};
 }