+++ examples/realworld.svelte.dev/src/routes/settings/index.svelte
@@ -8,7 +8,7 @@
 
 
 <script>
-	import { goto, stores } from '$app/client.js';
+	import { goto } from '$app/navigation';
 	import ListErrors from '$components/ListErrors.svelte';
 	import SettingsForm from './_SettingsForm.svelte';
 	import { post } from '$common/utils.js';