+++ packages/adapter-netlify/index.js
@@ -1,15 +1,10 @@
+'use strict';
+
 const fs = require('fs');
 const path = require('path');
 const toml = require('toml');
-const glob = require('tiny-glob/sync');
-const { prerender, generate_manifest_module } = require('@sveltejs/app-utils/renderer');
-const { mkdirp } = require('@sveltejs/app-utils/files');
 
-module.exports = async function builder({
-	dir,
-	manifest,
-	log
-}) {
+module.exports = async function adapter(builder) {
 	let netlify_config;
 
 	if (fs.existsSync('netlify.toml')) {
@@ -20,64 +15,34 @@ module.exports = async function builder({
 			throw err;
 		}
 	} else {
-		throw new Error('Missing a netlify.toml file. Consult https://github.com/sveltejs/kit/tree/master/packages/adapter-netlify#configuration');
+		// TODO offer to create one?
+		throw new Error(
+			'Missing a netlify.toml file. Consult https://github.com/sveltejs/kit/tree/master/packages/adapter-netlify#configuration'
+		);
 	}
 
 	if (!netlify_config.build || !netlify_config.build.publish || !netlify_config.build.functions) {
-		throw new Error('You must specify build.publish and build.functions in netlify.toml. Consult https://github.com/sveltejs/adapter-netlify#configuration');
+		throw new Error(
+			'You must specify build.publish and build.functions in netlify.toml. Consult https://github.com/sveltejs/adapter-netlify#configuration'
+		);
 	}
 
 	const publish = path.resolve(netlify_config.build.publish);
 	const functions = path.resolve(netlify_config.build.functions);
 
-	mkdirp(`${publish}/_app`);
-	mkdirp(`${functions}/render`);
-
-	// copy everything in `static`
-	glob('**/*', { cwd: 'static', filesOnly: true }).forEach(file => {
-		mkdirp(path.dirname(`${publish}/${file}`));
-		fs.copyFileSync(`static/${file}`, `${publish}/${file}`);
-	});
-
-	// copy client code
-	const client_code = path.resolve(dir, 'client');
-	glob('**/*', { cwd: client_code, filesOnly: true }).forEach(file => {
-		if (file[0] !== '.') {
-			mkdirp(path.dirname(`${publish}/_app/${file}`));
-			fs.copyFileSync(`${client_code}/${file}`, `${publish}/_app/${file}`);
-		}
-	});
-
-	// prerender
-	log.minor('Prerendering static pages...');
-	await prerender({
-		dir,
-		out: publish,
-		manifest,
-		log
-	});
-
-	// copy server code
-	const server_code = path.resolve(dir, 'server');
-	glob('**/*', { cwd: server_code, filesOnly: true }).forEach(file => {
-		if (file[0] !== '.') {
-			mkdirp(path.dirname(`${functions}/render/${file}`));
-			fs.copyFileSync(`${server_code}/${file}`, `${functions}/render/${file}`);
-		}
-	});
+	builder.copy_static_files(publish);
+	builder.copy_client_files(`${publish}/_app`);
+	builder.copy_server_files(`${functions}/render`);
 
 	// copy the renderer
-	fs.copyFileSync(path.resolve(__dirname, 'render.js'), `${functions}/render/index.js`);
-
-  // write manifest
-	fs.writeFileSync(`${functions}/render/manifest.js`, generate_manifest_module(manifest));
-
-	// copy client manifest
-	fs.copyFileSync(`${dir}/client.json`, `${functions}/render/client.json`);
-
-	// copy template
-	fs.writeFileSync(`${functions}/render/template.js`, `module.exports = ${JSON.stringify(fs.readFileSync('src/app.html', 'utf-8'))};`);
+	fs.copyFileSync(path.resolve(__dirname, 'files/render.js'), `${functions}/render/index.js`);
 
 	// create _redirects
 	fs.writeFileSync(`${publish}/_redirects`, '/* /.netlify/functions/render 200');
+
+	// prerender
+	builder.log.info('Prerendering static pages...');
+	await builder.prerender({
+		dest: publish
+	});
 };