+++ packages/app-utils/index.d.ts
@@ -1,7 +1,0 @@
-export {
-	EndpointManifest,
-	PageManifest,
-	PageComponentManifest,
-	SetupModule,
-	SSRComponentModule
-} from './src/types';