+++ packages/app-utils/rollup.config.js
@@ -1,10 +1,9 @@
 import { nodeResolve } from '@rollup/plugin-node-resolve';
-import typescript from '@rollup/plugin-typescript';
 import pkg from './package.json';
 
 const input = {};
-Object.keys(pkg.exports).forEach(key => {
-	input[key.replace(/^\.\//, '')] = `src/${key}/index.ts`;
+Object.keys(pkg.exports).forEach((key) => {
+	input[key.replace(/^\.\//, '')] = `src/${key}/index.js`;
 });
 
 export default {
@@ -26,11 +25,7 @@ export default {
 		}
 	],
 	plugins: [
-		nodeResolve(),
-		typescript()
+		nodeResolve()
 	],
-	external: [
-		...require('module').builtinModules,
-		...Object.keys(pkg.dependencies)
-	]
+	external: [...require('module').builtinModules, ...Object.keys(pkg.dependencies)]
 };