+++ packages/app-utils/src/http/index.js
@@ -1,0 +1,1 @@
+export { get_body } from './get_body';