+++ packages/create-svelte/cli/ambient.d.ts
@@ -1,7 +1,0 @@
-declare module '*.gitignore';
-
-declare module 'gitignore-parser';
-
-declare module 'prompts/lib/index';
-
-declare module 'tiny-glob/sync';