+++ packages/kit/src/api/utils.ts
@@ -1,6 +1,0 @@
-import { resolve } from 'path';
-import { copy } from '@sveltejs/app-utils/files';
-
-export function copy_assets() {
-	copy(resolve(__dirname, `../assets`), '.svelte/main');
-}