+++ packages/kit/src/runtime/navigation/index.js
@@ -1,0 +1,4 @@
+export { default as goto } from './goto';
+export { default as prefetch } from './prefetch';
+export { default as prefetchRoutes } from './prefetchRoutes';
+export { default as start } from './start';